修改配置文件

+++ src/config.cpp
@@ -1,9 +1,10 @@
 #include "logger.h"
 #include "config.h"
 
+#include "nvs_flash.h"
 #include "Preferences.h"
 
-config::ConfigValue config::value;
+config::ConfigValue config::config;
 
 static Preferences prefs;
 
@@ -11,17 +12,58 @@ void config::setup()
 {
   logger::debugln("Config is starting...");
 
-  prefs.begin(CONFIG_NAME);
-  // TODO:
-  prefs.clear();
-  if (prefs.isKey(CONFIG_VALUE_NAME))
+  esp_err_t err = nvs_flash_init();
+  if (err == ESP_ERR_NVS_NO_FREE_PAGES || err == ESP_ERR_NVS_NEW_VERSION_FOUND)
   {
-    prefs.getBytes(CONFIG_VALUE_NAME, &value, sizeof(ConfigValue));
+    const esp_partition_t *partition = esp_partition_find_first(ESP_PARTITION_TYPE_DATA, ESP_PARTITION_SUBTYPE_DATA_NVS, NULL);
+    if (partition != NULL)
+    {
+      err = esp_partition_erase_range(partition, 0, partition->size);
+      if (!err)
+      {
+        err = nvs_flash_init();
+      }
+      else
+      {
+        logger::warnln("Config failed to format the broken NVS partition!");
+      }
+    }
+    else
+    {
+      logger::warnln("Config could not find NVS partition");
+    }
+  }
+  if (err)
+  {
+    logger::warnln("Config failed to initialize NVS! Error: %u", err);
+  }
+
+  prefs.begin(CONFIG_NAME);
+
+  if (!prefs.isKey(CONFIG_VERSION_NAME))
+  {
+    // 不存则新建配置
+    prefs.putUShort(CONFIG_VERSION_NAME, CONFIG_VERSION_VALUE);
+    prefs.putBytes(CONFIG_DATA_NAME, &config, sizeof(ConfigValue));
   }
   else
   {
-    prefs.putBytes(CONFIG_VALUE_NAME, &value, sizeof(ConfigValue));
+    if (prefs.getUShort(CONFIG_VERSION_NAME) != CONFIG_VERSION_VALUE)
+    {
+      // 版本不一致重置配置
+      prefs.clear();
+      prefs.putUShort(CONFIG_VERSION_NAME, CONFIG_VERSION_VALUE);
+      prefs.putBytes(CONFIG_DATA_NAME, &config, sizeof(ConfigValue));
+    }
   }
 
+  // 读取配置
+  prefs.getBytes(CONFIG_DATA_NAME, &config, sizeof(ConfigValue));
+
   logger::debugln("Config is started.");
+}
+
+void config::save()
+{
+  prefs.putBytes(CONFIG_DATA_NAME, &config, sizeof(ConfigValue));
 }
